Auto cache book (via SW); remove keep UX
The keep UX sent a promise of permanence that browsers
are currently incapable of--as they only cache things...
and that cache can be cleared (intentionally by the user
or at the option of the browser to save space).

This now leans on Web App Manifest UX (if present) to
provide the installation and keeping feel. However, it
is still unclear whether the cache is anything like
permanent or at all within the users ultimate control.